Make some configuration values optional
When not using the non_caddyfile_config, it's kinda messy to have all the values kicking around - this just makes them optional.

If things fall back to the /etc/caddy file, then the "errors" look like the following. Not the clearest, but saves on a bunch of error checking logic in caddy.sh.
```
{"level":"info","ts":1608755727.2754512,"logger":"http","msg":"enabling automatic TLS certificate management","domains":["null"]}
```