Use `args` as an array not as a string
The `args` config accepts an array of strings. But when it uses it it treats it like a single string config by doing `${ARGS}`. It should be doing `${ARGS[@]}` in the run statement of Caddy so all elements of the args array are passed in as individual arguments to `caddy`.